added functions: -rm_most_recent_pdb_from_fasta_file() -read_fasta_pdb_ids() -write_index_file_from_file_path_list() -write_index_file_from_set()
modified function:
-create_fasta_file_from_index_file(),
which is called when there is no fasta file to append to.
added functionality in except block so that error during writing of vast amounts of data to a single fasta file will no longer require to rewrite the entire file.
It overwrites the original index file, so that in the next call of sync_pdb_db() the file can be appended according to the previously implemented index update rules.