-modified sequence selection to account for multiple underlying sequences being the same within the same alignment (aligning the same sequences multiple times at different points in the alignment file). for that the file red_aln_exact_match.index now has a dual index consisting of 1) the common sequence index and 2) a numbering index stating what place they hold in the list of alignment entries that have the same underlying sequence.
this still needs to be implemented in the downstream functions. (i.e. after function "create_job()")

-added function (?) "pstar_union()" to condense an entire alignment down to one score number

-modified function "import_sequence_list_from_fasta_aln()" to support fasta files that are NOT specifically MSAs.
It throws a warning, when sequences within a file don't all conform to the same length. Specifically: When AlignIO.read() throws a ValueError, we try using SeqIO.parse() to get sequences form fasta file. A warning is thrown no matter if SeqIO.parse() can read the file or not.